trivia-database-generation/requirements.txt
psycopg2-binary==2.9.10
python-dotenv==1.1.1
schema==0.7.7
